[direct] move asyncShift
